+++ protocols/dcutr/tests/lib.rs
@@ -72,26 +72,6 @@ async fn connect() {
 
     src.dial_and_wait(dst_relayed_addr.clone()).await;
 
-    loop {
-        match src
-            .next_swarm_event()
-            .await
-            .try_into_behaviour_event()
-            .unwrap()
-        {
-            ClientEvent::Dcutr(dcutr::Event::RemoteInitiatedDirectConnectionUpgrade {
-                remote_peer_id,
-                remote_relayed_addr,
-            }) => {
-                if remote_peer_id == dst_peer_id && remote_relayed_addr == dst_relayed_addr {
-                    break;
-                }
-            }
-            ClientEvent::Identify(_) => {}
-            other => panic!("Unexpected event: {other:?}."),
-        }
-    }
-
     let dst_addr = dst_tcp_addr.with(Protocol::P2p(dst_peer_id));
 
     let established_conn_id = src
@@ -107,9 +87,10 @@ async fn connect() {
 
     let reported_conn_id = src
         .wait(move |e| match e {
-            SwarmEvent::Behaviour(ClientEvent::Dcutr(
-                dcutr::Event::DirectConnectionUpgradeSucceeded { connection_id, .. },
-            )) => Some(connection_id),
+            SwarmEvent::Behaviour(ClientEvent::Dcutr(dcutr::Event {
+                result: Ok(connection_id),
+                ..
+            })) => Some(connection_id),
             _ => None,
         })
         .await;
@@ -218,6 +199,7 @@ async fn wait_for_reservation(
                 addr_observed = true;
             }
             SwarmEvent::Behaviour(ClientEvent::Identify(_)) => {}
+            SwarmEvent::NewExternalAddrCandidate { .. } => {}
             e => panic!("{e:?}"),
         }
     }