+++ protocols/relay/src/protocol/inbound_hop.rs
@@ -35,25 +35,11 @@ use crate::proto::message_v2::pb::mod_HopMessage::Type;
 use crate::protocol::MAX_MESSAGE_SIZE;
 
 #[derive(Debug, Error)]
-pub enum UpgradeError {
-    #[error("Fatal")]
-    Fatal(#[from] FatalUpgradeError),
-}
-
-impl From<quick_protobuf_codec::Error> for UpgradeError {
-    fn from(error: quick_protobuf_codec::Error) -> Self {
-        Self::Fatal(error.into())
-    }
-}
-
-#[derive(Debug, Error)]
-pub enum FatalUpgradeError {
+pub enum Error {
     #[error(transparent)]
     Codec(#[from] quick_protobuf_codec::Error),
     #[error("Stream closed")]
     StreamClosed,
-    #[error("Failed to parse response type field.")]
-    ParseTypeField,
     #[error("Failed to parse peer id.")]
     ParsePeerId,
     #[error("Expected 'peer' field to be set.")]
@@ -70,7 +56,7 @@ pub struct ReservationReq {
 }
 
 impl ReservationReq {
-    pub async fn accept(self, addrs: Vec<Multiaddr>) -> Result<(), FatalUpgradeError> {
+    pub async fn accept(self, addrs: Vec<Multiaddr>) -> Result<(), Error> {
         if addrs.is_empty() {
             tracing::debug!(
                 "Accepting relay reservation without providing external addresses of local node. \
@@ -104,7 +90,7 @@ impl ReservationReq {
         self.send(msg).await
     }
 
-    pub async fn deny(self, status: proto::Status) -> Result<(), FatalUpgradeError> {
+    pub async fn deny(self, status: proto::Status) -> Result<(), Error> {
         let msg = proto::HopMessage {
             type_pb: proto::HopMessageType::STATUS,
             peer: None,
@@ -116,7 +102,7 @@ impl ReservationReq {
         self.send(msg).await
     }
 
-    async fn send(mut self, msg: proto::HopMessage) -> Result<(), FatalUpgradeError> {
+    async fn send(mut self, msg: proto::HopMessage) -> Result<(), Error> {
         self.substream.send(msg).await?;
         self.substream.flush().await?;
         self.substream.close().await?;
@@ -135,7 +121,7 @@ impl CircuitReq {
         self.dst
     }
 
-    pub async fn accept(mut self) -> Result<(Stream, Bytes), FatalUpgradeError> {
+    pub async fn accept(mut self) -> Result<(Stream, Bytes), Error> {
         let msg = proto::HopMessage {
             type_pb: proto::HopMessageType::STATUS,
             peer: None,
@@ -160,7 +146,7 @@ impl CircuitReq {
         Ok((io, read_buffer.freeze()))
     }
 
-    pub async fn deny(mut self, status: proto::Status) -> Result<(), FatalUpgradeError> {
+    pub async fn deny(mut self, status: proto::Status) -> Result<(), Error> {
         let msg = proto::HopMessage {
             type_pb: proto::HopMessageType::STATUS,
             peer: None,
@@ -185,13 +171,13 @@ pub(crate) async fn handle_inbound_request(
     reservation_duration: Duration,
     max_circuit_duration: Duration,
     max_circuit_bytes: u64,
-) -> Result<Either<ReservationReq, CircuitReq>, FatalUpgradeError> {
+) -> Result<Either<ReservationReq, CircuitReq>, Error> {
     let mut substream = Framed::new(io, quick_protobuf_codec::Codec::new(MAX_MESSAGE_SIZE));
 
     let res = substream.next().await;
 
     if let None | Some(Err(_)) = res {
-        return Err(FatalUpgradeError::StreamClosed);
+        return Err(Error::StreamClosed);
     }
 
     let proto::HopMessage {
@@ -212,17 +198,17 @@ pub(crate) async fn handle_inbound_request(
         Type::CONNECT => {
             let peer_id_res = match peer {
                 Some(r) => PeerId::from_bytes(&r.id),
-                None => return Err(FatalUpgradeError::MissingPeer),
+                None => return Err(Error::MissingPeer),
             };
 
             let dst = match peer_id_res {
                 Ok(res) => res,
-                Err(_) => return Err(FatalUpgradeError::ParsePeerId),
+                Err(_) => return Err(Error::ParsePeerId),
             };
 
             Either::Right(CircuitReq { dst, substream })
         }
-        Type::STATUS => return Err(FatalUpgradeError::UnexpectedTypeStatus),
+        Type::STATUS => return Err(Error::UnexpectedTypeStatus),
     };
 
     Ok(req)